ci: fix goreleaser relase tag choice (#35)
If a pre-release tag (e.g. `v0.0.5-test5`) and a "final" release tag
(`v0.0.5`) point to the same commit hash, goreleaser will by default
choose the pre-release tag when computing release parameters.

This should hopefully cause it to choose the most recently created tag,
and we assume that will be the final release tag.